# contracts/openapi/control-plane.v3.yaml
openapi: 3.1.0
info:
  title: NeKiro Control Plane API
  version: 3.0.0
  description: >-
    Gateway-owned Northbound API. Invocation v2 returns transient Agent output
    on the invocation request and exposes only metadata through Ledger reads.
servers:
  - url: https://api.nekiro.dev
    description: NeKiro Gateway destination
paths:
...
  /v3/workspaces:
    post:
      operationId: createWorkspace
      summary: Create an owner-controlled Workspace
      description: >-
        Creates one logical authorization and audit boundary. The authenticated
        caller becomes the immutable owner. Owner identity and timestamps are
        never accepted from request data. Repeating an existing workspaceId is
        a conflict rather than an idempotent success.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        x-nekiro-max-body-bytes: 1048576
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          headers:
            x-nek-trace-id:
              $ref: "#/components/headers/TraceID"
          content:
            application/json:
              schema:
                $ref: ../schemas/workspace.v1.schema.json
        "400":
          $ref: "#/components/responses/WorkspaceValidationError"
        "401":
          $ref: "#/components/responses/WorkspaceUnauthenticated"
        "409":
          $ref: "#/components/responses/WorkspaceConflict"
        "503":
          $ref: "#/components/responses/WorkspaceDependencyError"
  /v3/workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      summary: Read an owned Workspace
      description: >-
        Returns the exact durable Workspace only to its immutable owner.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Exact Workspace
          headers:
            x-nek-trace-id:
              $ref: "#/components/headers/TraceID"
          content:
            application/json:
              schema:
                $ref: ../schemas/workspace.v1.schema.json
        "400":
          $ref: "#/components/responses/WorkspaceValidationError"
        "401":
          $ref: "#/components/responses/WorkspaceUnauthenticated"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
        "503":
          $ref: "#/components/responses/WorkspaceDependencyError"
  /v3/workspaces/{workspaceId}/installations:
    post:
      operationId: installAgent
      summary: Install and pin one published Agent version
      description: >-
        Selects the highest published version satisfying the exact submitted
        SemVer constraint, then requires that exact version to have a published
        trusted Release (or an explicitly marked pre-v4 legacy marker),
        validates the accepted permission subset against that Card, and creates
        one enabled immutable pin. It does not silently downgrade when the
        selected version fails the Release gate.
        A current enabled or disabled Installation for the same Agent conflicts.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        x-nekiro-max-body-bytes: 1048576
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallAgentRequest"
      responses:
        "201":
          description: Agent installed in Workspace
          headers:
            x-nek-trace-id:
              $ref: "#/components/headers/TraceID"
          content:
            application/json:
              schema:
                $ref: ../schemas/installation.v2.schema.json
        "400":
          $ref: "#/components/responses/WorkspaceValidationError"
        "401":
          $ref: "#/components/responses/WorkspaceUnauthenticated"
        "403":
          $ref: "#/components/responses/WorkspaceInstallForbidden"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
        "409":
          $ref: "#/components/responses/WorkspaceConflict"
        "503":
          $ref: "#/components/responses/WorkspaceDependencyError"
    get:
      operationId: listInstallations
      summary: List current and historical Installations
      description: >-
        Returns at most the requested bounded page of enabled, disabled, and
        uninstalled Installations ordered by installedAt and then
        installationId ascending. The opaque cursor is bound to the Workspace
        and page size. An existing owned Workspace with no records returns an
        explicit empty items array without a cursor.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/InstallationLimit"
        - $ref: "#/components/parameters/InstallationCursor"
      responses:
        "200":
          description: Bounded page of current and historical Installations
          headers:
            x-nek-trace-id:
              $ref: "#/components/headers/TraceID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallationList"
        "400":
          $ref: "#/components/responses/WorkspaceValidationError"
        "401":
          $ref: "#/components/responses/WorkspaceUnauthenticated"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
        "503":
          $ref: "#/components/responses/WorkspaceDependencyError"
  /v3/workspaces/{workspaceId}/installations/{installationId}:
    get:
      operationId: getInstallation
      summary: Read one current or historical Installation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/InstallationId"
      responses:
        "200":
          description: Exact Installation
          headers:
            x-nek-trace-id:
              $ref: "#/components/headers/TraceID"
          content:
            application/json:
              schema:
                $ref: ../schemas/installation.v2.schema.json
        "400":
          $ref: "#/components/responses/WorkspaceValidationError"
        "401":
          $ref: "#/components/responses/WorkspaceUnauthenticated"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
        "503":
          $ref: "#/components/responses/WorkspaceDependencyError"
    patch:
      operationId: updateInstallation
      summary: Enable or disable a current Installation
      description: >-
        Applies exactly one enabled-to-disabled or disabled-to-enabled
        transition. Same-state requests and every transition from uninstalled
        conflict rather than returning idempotent success.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/InstallationId"
      requestBody:
        required: true
        x-nekiro-max-body-bytes: 1048576
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstallationRequest"
      responses:
        "200":
          description: Updated installation
          headers:
            x-nek-trace-id:
              $ref: "#/components/headers/TraceID"
          content:
            application/json:
              schema:
                $ref: ../schemas/installation.v2.schema.json
        "400":
          $ref: "#/components/responses/WorkspaceValidationError"
        "401":
          $ref: "#/components/responses/WorkspaceUnauthenticated"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
        "409":
          $ref: "#/components/responses/WorkspaceConflict"
        "503":
          $ref: "#/components/responses/WorkspaceDependencyError"
    delete:
      operationId: uninstallAgent
      summary: Uninstall and preserve one disabled Installation
      description: >-
        Transitions only a disabled Installation to uninstalled and returns the
        preserved terminal fact. Enabled and already-uninstalled records
        conflict; uninstall is not idempotent.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/InstallationId"
      responses:
        "200":
          description: Preserved terminal Installation
          headers:
            x-nek-trace-id:
              $ref: "#/components/headers/TraceID"
          content:
            application/json:
              schema:
                $ref: ../schemas/installation.v2.schema.json
        "400":
          $ref: "#/components/responses/WorkspaceValidationError"
        "401":
          $ref: "#/components/responses/WorkspaceUnauthenticated"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
        "409":
          $ref: "#/components/responses/WorkspaceConflict"
        "503":
          $ref: "#/components/responses/WorkspaceDependencyError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Gateway-verified bearer credential mapped to one trusted caller ID.
        Public caller-ID headers are not an authentication source.
  headers:
    TraceID:
      description: Gateway-assigned request trace identifier.
      required: true
      schema:
        $ref: ../schemas/common.v1.schema.json#/$defs/traceId
  parameters:
...
    WorkspaceId:
      in: path
      name: workspaceId
      required: true
      schema:
        $ref: ../schemas/common.v1.schema.json#/$defs/workspaceId
    InstallationId:
      in: path
      name: installationId
      required: true
      schema:
        $ref: ../schemas/common.v1.schema.json#/$defs/installationId
    InstallationLimit:
      in: query
      name: limit
      required: true
      description: >-
        Required maximum number of Installations in one page. Values outside
        1-100 are validation failures.
      schema:
        type: integer
        minimum: 1
        maximum: 100
    InstallationCursor:
      in: query
      name: cursor
      required: false
      description: >-
        Opaque continuation bound to this Workspace, the page size, and the
        last installedAt/installationId ordering tuple. Malformed or mismatched
        cursors are validation failures and never restart traversal.
      schema:
        type: string
        minLength: 1
...
  responses:
...
    WorkspaceValidationError:
      description: Workspace request validation failed
      headers:
        x-nek-trace-id:
          $ref: "#/components/headers/TraceID"
      x-platform-error-codes: [VALIDATION_ERROR]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v3.schema.json
    WorkspaceUnauthenticated:
      description: A valid Gateway bearer identity is required
      headers:
        x-nek-trace-id:
          $ref: "#/components/headers/TraceID"
      x-platform-error-codes: [UNAUTHENTICATED]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v3.schema.json
    WorkspaceForbidden:
      description: The authenticated caller does not own the Workspace
      headers:
        x-nek-trace-id:
          $ref: "#/components/headers/TraceID"
      x-platform-error-codes: [FORBIDDEN]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v3.schema.json
    WorkspaceInstallForbidden:
      description: The caller is forbidden or no matching published Release is installable
      headers:
        x-nek-trace-id:
          $ref: "#/components/headers/TraceID"
      x-platform-error-codes: [FORBIDDEN, AGENT_RELEASE_UNPUBLISHED, AGENT_RELEASE_SUSPENDED, AGENT_RELEASE_REVOKED]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v3.schema.json
    WorkspaceNotFound:
      description: The requested Workspace, Installation, or published install candidate was not found
      headers:
        x-nek-trace-id:
          $ref: "#/components/headers/TraceID"
      x-platform-error-codes: [NOT_FOUND]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v3.schema.json
    WorkspaceConflict:
      description: The Workspace or current Installation already exists, or the lifecycle transition is illegal
      headers:
        x-nek-trace-id:
          $ref: "#/components/headers/TraceID"
      x-platform-error-codes: [CONFLICT]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v3.schema.json
    WorkspaceDependencyError:
      description: A required Workspace, Catalog, or persistence dependency failed
      headers:
        x-nek-trace-id:
          $ref: "#/components/headers/TraceID"
      x-platform-error-codes: [DEPENDENCY_ERROR]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v3.schema.json
...
  schemas:
    CreateWorkspaceRequest:
      type: object
      additionalProperties: false
      required: [workspaceId]
      properties:
        workspaceId:
          $ref: ../schemas/common.v1.schema.json#/$defs/workspaceId
    InstallAgentRequest:
      type: object
      additionalProperties: false
      required: [agentId, versionConstraint, acceptedPermissions]
      properties:
        agentId:
          $ref: ../schemas/common.v1.schema.json#/$defs/agentId
        versionConstraint:
          $ref: ../schemas/common.v1.schema.json#/$defs/semverRange
        acceptedPermissions:
          type: array
          uniqueItems: true
          items:
            $ref: ../schemas/common.v1.schema.json#/$defs/permissionId
...